Add Slack reply with Block kit messages
This adds the capability to JiraUnfurler and DomainUnfurler in general
to now reply to the slack message.

The message is formatted as a block kit message. The models for the
Slack Block Kit are adapted from safir.slack.blockkit, but had to be
reimplemented to more generic. At the moment we support the text section
and the context section with the general Slack text object.