Stop using quotes for main content
The original intent of the mrkdwn quote was to emulate the attachments
API's ability to add a coloured bar to content blocks. We don't want to
use attachments because the API is technically deprecated. However,
we've also found that the mrkdwn quote wraps poorly on Slack mobile
views and has done so for many months now and is unlikely to be fixed
upstream. To mitigate this we'll simply remove the quote from
unfurlbot's messages.