Update dependencies for new Squarebot client
Update Python and pre-commit dependencies to pick up the new
Squarebot client, since it changed the Faststream model for the
Kafka messages corresponding to Slack messages from bot users. The
mismatch was causing bot messages to not be processed.

Update the `Makefile` rules to use universal dependency files and
to remove the `update-deps-no-hashes` target, since it's no longer
necessary with current uv.