Stop using the KafkaRouter lifespan_context
Previously, the lifespan function for the FastAPI app was calling
the `lifespan_context` attribute of the `KafkaRouter` inside an
`async with` block. Current FastStream documentation does not
document this pattern and says to simply include the router into
the FastAPI app, and in practice including this `async with` call
with the latest Python dependencies creates a bunch of asyncio
tasks that are never awaited. Remove this pattern, which seems to
fix the warnings and still passes tests.