Upgrade to the latest FastStream
The older version of FastStream that this package had pinned was
incompatible with the latest version of FastAPI. Update FastStream,
which requires reworking `ConsumerContextDependency` so that it uses
the dependency syntax to get the raw Kafka message. This also requires
handling the case of multiple messages for typing purposes.

This uncovered problems with FastAPI type resolution that were fixed
by removing `from __future__ import annotations` from handler and
dependency modules and letting the type resolution happen at import
time.