Add clear logging for unfurling a Jira issue
This will make it easier to track when what keys are being unfurled and
from what message they were triggered. The info message contains what
key was unfurled and what channel/thread it was sent to. Also includes
the kafka message context to understand if the same message is being
processed multiple times.